+++ src/main/java/com/feelmycode/parabole/controller/testController.java
@@ -1,0 +1,5 @@
+package com.feelmycode.parabole.controller;
+
+public class testController {
+
+}
