+++ src/main/java/com/feelmycode/parabole/domain/TestDomain.java
@@ -1,0 +1,4 @@
+package com.feelmycode.parabole.domain;
+
+public class TestDomain {
+}
