+++ src/main/java/com/feelmycode/parabole/global/config/TestConfig.java
@@ -1,0 +1,4 @@
+package com.feelmycode.parabole.global.config;
+
+public class TestConfig {
+}
