+++ src/main/java/com/feelmycode/parabole/global/error/TestError.java
@@ -1,0 +1,4 @@
+package com.feelmycode.parabole.global.error;
+
+public class TestError {
+}
