+++ src/main/java/com/feelmycode/parabole/global/interceptor/TestInterceptor.java
@@ -1,0 +1,4 @@
+package com.feelmycode.parabole.global.interceptor;
+
+public class TestInterceptor {
+}
