+++ src/main/java/com/feelmycode/parabole/global/jwt/TestJwt.java
@@ -1,0 +1,4 @@
+package com.feelmycode.parabole.global.jwt;
+
+public class TestJwt {
+}
