+++ src/main/java/com/feelmycode/parabole/global/resolver/TestResolver.java
@@ -1,0 +1,4 @@
+package com.feelmycode.parabole.global.resolver;
+
+public class TestResolver {
+}
