+++ src/main/java/com/feelmycode/parabole/global/util/TestUtil.java
@@ -1,0 +1,4 @@
+package com.feelmycode.parabole.global.util;
+
+public class TestUtil {
+}
