+++ src/main/java/com/feelmycode/parabole/infra/TestInfra.java
@@ -1,0 +1,4 @@
+package com.feelmycode.parabole.infra;
+
+public class TestInfra {
+}
