+++ src/main/java/com/feelmycode/parabole/repository/TestRespostory.java
@@ -1,0 +1,5 @@
+package com.feelmycode.parabole.repository;
+
+public class TestRespostory {
+
+}
