+++ src/main/java/com/feelmycode/parabole/web/TestWeb.java
@@ -1,0 +1,4 @@
+package com.feelmycode.parabole.web;
+
+public class TestWeb {
+}
