--- src/main/java/com/fellmycode/parabole/controller/testController.java
@@ -1,5 +1,0 @@
-package com.fellmycode.parabole.controller;
-
-public class testController {
-
-}
