--- src/main/java/com/fellmycode/parabole/domain/TestDomain.java
@@ -1,4 +1,0 @@
-package com.fellmycode.parabole.domain;
-
-public class TestDomain {
-}
