--- src/main/java/com/fellmycode/parabole/global/config/TestConfig.java
@@ -1,4 +1,0 @@
-package com.fellmycode.parabole.global.config;
-
-public class TestConfig {
-}
