--- src/main/java/com/fellmycode/parabole/global/error/TestError.java
@@ -1,4 +1,0 @@
-package com.fellmycode.parabole.global.error;
-
-public class TestError {
-}
