--- src/main/java/com/fellmycode/parabole/global/interceptor/TestInterceptor.java
@@ -1,4 +1,0 @@
-package com.fellmycode.parabole.global.interceptor;
-
-public class TestInterceptor {
-}
