--- src/main/java/com/fellmycode/parabole/global/jwt/TestJwt.java
@@ -1,4 +1,0 @@
-package com.fellmycode.parabole.global.jwt;
-
-public class TestJwt {
-}
