--- src/main/java/com/fellmycode/parabole/global/resolver/TestResolver.java
@@ -1,4 +1,0 @@
-package com.fellmycode.parabole.global.resolver;
-
-public class TestResolver {
-}
