--- src/main/java/com/fellmycode/parabole/global/util/TestUtil.java
@@ -1,4 +1,0 @@
-package com.fellmycode.parabole.global.util;
-
-public class TestUtil {
-}
