--- src/main/java/com/fellmycode/parabole/infra/TestInfra.java
@@ -1,4 +1,0 @@
-package com.fellmycode.parabole.infra;
-
-public class TestInfra {
-}
