--- src/main/java/com/fellmycode/parabole/repository/TestRespostory.java
@@ -1,5 +1,0 @@
-package com.fellmycode.parabole.repository;
-
-public class TestRespostory {
-
-}
