--- src/main/java/com/fellmycode/parabole/service/testService.java
@@ -1,5 +1,0 @@
-package com.fellmycode.parabole.service;
-
-public class testService {
-
-}
