--- src/main/java/com/fellmycode/parabole/web/TestWeb.java
@@ -1,4 +1,0 @@
-package com.fellmycode.parabole.web;
-
-public class TestWeb {
-}
