+++ src/test/java/com/feelmycode/parabole/ParaboleApplicationTests.java
@@ -1,4 +1,4 @@
-package com.fellmycode.parabole;
+package com.feelmycode.parabole;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
